--- src/pages/CRUD.jsx
@@ -1,8 +1,0 @@
-import React, {useState}from 'react'
-
-
-function CRUD() {
-     <h2>virker det eller hvad llcrud</h2>
-}
-
-export default CRUD;
