+++ src/pages/Landing.jsx
@@ -4,52 +4,14 @@ import {USER_ENDPOINT, SCHOOL_ENDPOINT} from '../settings.js';
 
 export default function Landing() {
   // const[id,idchange]=useState("");
-  const [username, setUserName] = useState("");
+  /*const [username, setUserName] = useState("");
   const [name, setName] = useState("");
   const [password, setPassword] = useState("");
   const [phone, setPhone] = useState("");
   const [address, setAddress] = useState("");
   const [zipcode, setZipcode] = useState("");
-  const [schoolId, setSchoolId] = useState("");
+  const [schoolId, setSchoolId] = useState("");*/
 
-  const [schoolName, setSchoolName] = useState("");
-  const [location, setLocation] = useState("");
-
-  const [data, setData] = useState(null);
-
-  
-  useEffect(() => {
-    fetch(SCHOOL_ENDPOINT)
-      .then((res) => {
-        return res.json()
-      })
-      .then(data => {
-        setData(data)
-    })
-      .catch((err) => {
-        console.log(err.message);
-      });
-  }, []);
-
-
-
-  const handlesubmit2 = (e) => {
-    e.preventDefault();
-    const schooldata = {schoolName, location};
-
-
-    fetch(SCHOOL_ENDPOINT, {
-      method: "POST",
-      headers: {"content-type": "application/json"},
-      body: JSON.stringify(schooldata)
-    }).then((res) => {
-      alert('Saved successfully.')
-    }).catch((err) => {
-      console.log(err.message)
-    }, [])
-  }
-
-  
   const [fields, setFields] = useState({});
 
   useEffect(() => {
@@ -159,75 +121,6 @@ export default function Landing() {
               </div>
             </div>
           </form>
-        </div>
-      </section>
-
-      <section className="row">
-        <div className="col-8">
-          <form className="container" onSubmit={handlesubmit2}>
-            <div className="card" style={{"textAlign": "left"}}>
-              <div className="card-title">
-                <h2>Create School</h2>
-              </div>
-              <div className="card-body">
-
-                <div className="row">
-                  
-                  <div className="col-lg-12">
-                    <div className="form-group">
-                      <label>School</label>
-                      <input onChange={e => setSchoolName(e.target.value)} className="form-control"></input>
-                    </div>
-                  </div>
-
-                  <div className="col-lg-12">
-                    <div className="form-group">
-                      <label>Address</label>
-                      <input onChange={e => setLocation(e.target.value)} className="form-control"></input>
-                    </div>
-                  </div>
-
-                  <div className="col-lg-12">
-                    <div className="form-group">
-                      <button className="btn btn-primary w-20"  type="submit">Save</button>
-                    </div>
-                  </div>
-
-                </div>
-
-              </div>
-            </div>
-          </form>
-        </div>
-      </section>
-
-      <section className="row">
-        <div className="card">
-          <div className="card-title">
-              <h2>List of Schools</h2>
-          </div>
-          <div className="card-body">
-
-            <table className="table table-default">
-              <thead className="bg-dark text-white">
-                <tr>
-                  <td>School</td>
-                  <td>Address</td>
-                </tr>
-              </thead>
-              <tbody>
-                {data?.length && data.map((school) => {
-                  return (
-                    <tr key={school.id}>
-                    <td>{school.schoolName}</td>
-                    <td>{school.location}</td>
-                    </tr>      
-                  );
-                })}
-              </tbody>
-            </table>
-
-          </div>
         </div>
       </section>
     </main>
